chore: renamed cypher dir to repositories
