+++ apps/web/src/graphql/repositories/type/getInterfaceFieldByKey.cypher
