+++ apps/web/src/graphql/repositories/type/getTypeById.cypher
