+++ apps/web/src/graphql/repositories/type/isTypeDescendantOf.cypher
