+++ apps/web/src/graphql/resolvers/mutation/field/deleteFieldEdge.ts
@@ -1,8 +1,8 @@
 import { DeleteFieldInput } from '@codelab/shared/abstract/codegen-v2'
 import { IFieldResolver } from '@graphql-tools/utils/Interfaces'
 import { RxTransaction } from 'neo4j-driver'
-import { DisconnectFieldResponse, typeRepository } from '../../../cypher'
 import { getDriver } from '../../../infra/driver'
+import { DisconnectFieldResponse, typeRepository } from '../../../repositories'
 
 const driver = getDriver()
